- закончил с графиком всех расходов, теперь работает нормально  - добавил кнопку добавления нового расхода, в случае если расходы уже есть  - пофиксил баг с тем что окно добавления расхода открывается только один раз, дело было с переменной hangeStatus, по значению которой мы узнавали что данные о расходе записанны и пора вернутся на предыдущий фрагмент. Нужно было после изменения значения hangeStatus, задавть значения null. Ну я так сделал  - залил новые дефолтные иконки  - отображение фоток в фрагменте добавления расхода теперь нормально (scalrType и maxHeight - спасибо вам)  - изменил фотку загрузки изображения  - добавил проверку что если сегодняшнея дата находится в длиапазоне существующего путешествия - то переходим на него  - еще пофиксил мелкие баги
 Что не сделал: не правильно работает PersonView, всегда имеет тип adult, не грузятся фотки путешествий на фрагменте YouTravels, падение после добавления нового путешествия...

+++ app/src/main/java/com/twosmalpixels/travel_notes/views/shedule/CategoryExpenceData.kt
@@ -4,8 +4,8 @@ class CategoryExpenceData(
     val image: Int,
     var height: Float,
     val name: Int,
-    val mainCurrAmount: Int,
-    val additionalCurrAmount: Int,
+    val mainCurrAmount: Float,
+    val additionalCurrAmount: Float,
     val mainCurrencyIso: String,
     val additionalCurrencyIso: String
 ) {
